chore: improve folders structure
move files with types to corresponding component's folder; index.js as public API; set up absolute imports
