+++ src/components/todo-container/index.ts
@@ -1,0 +1,5 @@
+import TodoContainer from '@components/todo-container/todo-container';
+import { TodoContainerType } from '@components/todo-container/todo-container-types';
+
+export { TodoContainer };
+export type { TodoContainerType };
