+++ src/components/todo-container/todo-container-types.ts
@@ -1,0 +1,8 @@
+import { TodoType } from '@components/todo';
+
+export type TodoContainerType = {
+  id: string;
+  title: string;
+  active: boolean;
+  todos: TodoType[];
+};
