+++ src/components/todo-container/todo-container.tsx
@@ -1,14 +1,14 @@
-import './TodoContainer.scss';
+import './todo-container.scss';
 import React, { Dispatch, SetStateAction, useState } from 'react';
 import { useDrop } from 'react-dnd';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faTrashCan } from '@fortawesome/free-regular-svg-icons';
-import Items from '../types/types';
-import { TodoContainerType, TodoType } from '../../types/todoTypes';
-import Input from '../UI/Input/Input';
-import useAutoFocus from '../../hooks/useAutofocus';
-import CountIcon from '../UI/CountIcon/CountIcon';
-import Button from '../UI/Button/Button';
+import { Input } from '@components/ui/input';
+import useAutofocus from '@hooks/use-autofocus';
+import CountIcon from '@components/ui/count-icon/count-icon';
+import Button from '@components/ui/button/button';
+import { Items, TodoType } from '@components/todo';
+import { TodoContainerType } from './todo-container-types';
 
 type TodoContainerProps = {
   container: TodoContainerType;
@@ -77,7 +77,7 @@ export default function TodoContainer(props: TodoContainerProps): JSX.Element {
     });
   };
 
-  const titleRef = useAutoFocus(container.active, deactivateFocus);
+  const titleRef = useAutofocus(container.active, deactivateFocus);
 
   const handleSaveTitle = (id: TodoContainerType['id']) => {
     setTodoContainers((containers) => {
