+++ src/components/todo/index.ts
@@ -1,0 +1,5 @@
+import Todo from '@components/todo/todo';
+import { TodoType, Items } from '@components/todo/todo-types';
+
+export { Todo, Items };
+export type { TodoType };
