--- src/components/types/types.ts
@@ -1,5 +1,0 @@
-enum Items {
-  TODO = 'todo',
-}
-
-export default Items;
