+++ src/components/ui/count-icon/count-icon.scss
