+++ src/components/ui/count-icon/count-icon.tsx
@@ -1,4 +1,4 @@
-import './CountIcon.scss';
+import './count-icon.scss';
 
 type CountIconProps = {
   children: number | string;
