+++ src/components/ui/input/index.ts
@@ -1,0 +1,4 @@
+import Input from '@components/ui/input/input';
+import TextArea from '@components/ui/input/text-area';
+
+export { Input, TextArea };
